Add async consumer
